どうもシャムでーす 今日はオフ会当日なんですけど
