+++ P+P/Assets/Scripts/CameraCont.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03d3509d7bb1e9c40957570b237346b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
