+++ P+P/Assets/Scripts/SelfManagementOfTrackedDevices.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections;
+﻿
+/*
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using Valve.VR;
@@ -71,3 +73,4 @@ public class SelfManagementOfTrackedDevices : MonoBehaviour
         }
     }
 }
+*/
