--- P+P/Assets/ViveTracker_input.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1e1f2980a9b8af245b7641ebd4dc7382
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
